pyproject.toml
[tool.poetry]
name = "base"
version = "0.1.0"
description = "Base project for backend project"
authors = ["Dale Caringal <daraloca@live.it>"]


[tool.poetry.dependencies]
boto3 = "^1.34.103"
fastapi = "^0.111.0"
mypy-boto3-dynamodb = "^1.34.97"
psycopg2-binary = "^2.9.9"
pymongo = "^4.7.2"
python = "^3.12"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
requests = "^2.31.0"
structlog = "^24.1.0"
tenacity = "^8.3.0"
ulid-py = "^1.1.0"

# extras dependencies
pre-commit = { version = "^3.7.1", optional = true }
python-dotenv = { version = "^1.0.1", optional = true }
uvicorn = {version = "^0.29.0", optional = true}

[tool.poetry.dev-dependencies]
black = "^24.4.2"
factory-boy = "^3.3.0"
flake8 = "^7.0.0"
ipython = "^8.24.0"
isort = "^5.13.2"
pudb = "^2024.1"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
pytest-icdiff = "^0.9"
pytest-mock = "^3.14.0"

[tool.poetry.extras]
# local extras are used only on the local machine
local = ["pre-commit", "python-dotenv", "uvicorn"]

[tool.black]
line-length = 100
target-version = ['py312']
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.venv
  )/
)
'''

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
addopts = "--strict-markers"
markers = ["integration", "slow"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
